api/requirements.txt
Flask==2.0.2
PyYAML==6.0
gevent==21.12.0
gunicorn==20.1.0
PyMySQL==1.0.2
SQLAlchemy==1.4.29
kubernetes==21.7.0
pyhumps==3.5.0